driver finds dependencies and passes to worker
Driver looks for command line dependencies and passes their full paths
to the worker on each invocation in the worker's command
invocation. This eliminates searching for four dependencies on each
invocation of a worker.

When worker does not receive the path of one or more of its
dependencies, it looks for it, and if the dependency is not mdls,
which is optional, the worker bails when it cannot be found.